Separate schema into separate file
